Add clear input function to translation module
Introduced a "Clear" button and corresponding 'clearInput' function in TranslationModule component. This allows the user to clear the input and output content in the translation module. Also adjusted the button CSS class and re-ordered disable condition checks for better readability. Modified skeleton CSS to use em for height for better responsiveness.